Textexpr without first newline improved, empty statement possible, throw on missing main
Now the <::> can be used for a text start without newline written with ident designation
* usage: subtext(args) <::>textexpr<.> possible and proposed
* intern: ZbnfItemAttribute ?.skipFirstnl= and ?.ident= used.
* Some simplifications for inhibit repeat same programming:
** JZtxtcmdSyntax: definition of text::=
** JZtxtcmdScript: common use of JZcmditem#new_textExprStatements also in StatementList
   for that to obvious: rename JZcmditem#new_textExpr to JZcmditem#new_textExprStatements

Featurefix: Now empty statementlist is possible.
The problem for that was that an empty statement list such as main() { }
has had an obscure error message. Secondly sometime an empty statement block may be
* Change simple, statementBlock::= { [ <statement?> ] } with option braces.

JZtxtcmdExecuter: throw on missing main is a better error messaging.

ZbnfSyntaxPrescript: only formally changes this.